Refactor Modal
