+++ frontend/src/components/Tag.tsx
@@ -4,36 +4,148 @@ import { TagType } from '../utils/types';
 import React from 'react';
 import '../styles/Tag.scss';
 import { UniqueIdentifier } from '@dnd-kit/core';
+import { Fragment } from 'react';
+import { Menu, Transition } from '@headlessui/react';
 
-interface TagProps {
+interface previewTagProps {
+	tag: TagType;
+}
+
+interface tagProps {
 	tag: TagType;
 	onRemoveTag: (tagId: UniqueIdentifier) => void;
 }
 
-const Tag = ({ tag, onRemoveTag }: TagProps) => {
-	const [hiddenClose, setHiddenClose] = React.useState(true);
+interface tagEditorProps {
+	tags: TagType[];
+	currentTags: TagType[];
+	setCurrentTags: React.Dispatch<React.SetStateAction<TagType[]>>;
+	onAddTag: (tagName: string) => void;
+	onRemoveTag: (tagId: UniqueIdentifier) => void;
+	onSelectTag: (tag: TagType) => void;
+}
 
-	return (
-		<li id={tag.id.toString()}>
-			<div
-				className={clsx('tag', tag.color)}
-				onMouseEnter={() => {
-					setHiddenClose(false);
-				}}
-				onMouseLeave={() => {
-					setHiddenClose(true);
-				}}
-			>
-				<span className="tag-title">{tag.name}</span>
-				<span
-					className={clsx('tag-close-icon', tag.color, { hidden: hiddenClose })}
-					onClick={() => onRemoveTag(tag.id)}
-				>
-					<FaTrashAlt />
-				</span>
+class TagDisplay {
+	constructor() {
+		throw new Error('TagDisplay is a static class');
+	}
+
+	private static classNames = (...classes: any) => {
+		return classes.filter(Boolean).join(' ');
+	};
+
+	public static previewTag = ({ tag }: previewTagProps) => {
+		return (
+			<div id={tag.id.toString()}>
+				<div className={clsx('tag', tag.color)}>
+					<span className="tag-title">{tag.name}</span>
+				</div>
 			</div>
-		</li>
-	);
-};
+		);
+	};
+	public static tag = ({ tag, onRemoveTag }: tagProps) => {
+		const [hiddenClose, setHiddenClose] = React.useState(true);
 
-export default Tag;
+		return (
+			<div id={tag.id.toString()}>
+				<div
+					className={clsx('tag', tag.color)}
+					onMouseEnter={() => {
+						setHiddenClose(false);
+					}}
+					onMouseLeave={() => {
+						setHiddenClose(true);
+					}}
+				>
+					<span className="tag-title">{tag.name}</span>
+					<span
+						className={clsx('tag-close-icon', tag.color, { hidden: hiddenClose })}
+						onClick={() => onRemoveTag(tag.id)}
+					>
+						<FaTrashAlt />
+					</span>
+				</div>
+			</div>
+		);
+	};
+
+	public static tagsEditor = ({ tags, currentTags, onRemoveTag, onAddTag, onSelectTag }: tagEditorProps) => {
+		const [placeholder, setPlaceholder] = React.useState('Add / Edit');
+		const [menuVisible, setMenuVisible] = React.useState(false);
+
+		const getNoneSelectedTags = () => {
+			// Get all tags that are not selected (id)
+			return tags.filter((tag: TagType) => !currentTags.find((currentTag: TagType) => currentTag.id === tag.id));
+		};
+
+		const addTags = (event: any) => {
+			if (event.target.value !== '') {
+				if (tags.find((tag: TagType) => tag.name === event.target.value)) {
+					return;
+				}
+				onAddTag(event.target.value);
+				event.target.value = '';
+			}
+		};
+		return (
+			<div>
+				<div className="tags-input">
+					<ul id="tags">
+						{currentTags.map((tag: TagType) => (
+							<TagDisplay.tag key={tag.id} tag={tag} onRemoveTag={onRemoveTag} />
+						))}
+						<input
+							type="text"
+							onKeyUp={(event) => (event.key === 'Enter' ? addTags(event) : null)}
+							placeholder={placeholder}
+							onFocus={() => {
+								setPlaceholder('Press enter to add tag');
+								if (getNoneSelectedTags().length > 0) {
+									setMenuVisible(true);
+								}
+							}}
+							onBlur={() => {
+								setPlaceholder('Add / Edit');
+								setMenuVisible(false);
+							}}
+						/>
+					</ul>
+				</div>
+				<Menu>
+					<Transition
+						as={Fragment}
+						show={menuVisible}
+						enter="transition ease-out duration-100"
+						enterFrom="transform opacity-0 scale-95"
+						enterTo="transform opacity-100 scale-100"
+						leave="transition ease-in duration-75"
+						leaveFrom="transform opacity-100 scale-100"
+						leaveTo="transform opacity-0 scale-95"
+					>
+						<Menu.Items className="absolute z-10 origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
+							<div className="py-1 drop-down-tag">
+								{getNoneSelectedTags().map((tag: TagType) => (
+									<Menu.Item key={tag.id}>
+										{({ active }) => (
+											<div
+												className={this.classNames(
+													active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+													'block px-4 py-2 text-sm'
+												)}
+												onClick={() => onSelectTag(tag)}
+											>
+												{tag.name}
+											</div>
+										)}
+									</Menu.Item>
+								))}
+							</div>
+						</Menu.Items>
+					</Transition>
+				</Menu>
+			</div>
+		);
+	};
+}
+
+export default TagDisplay;
