+++ frontend/src/styles/Tag.scss
@@ -6,10 +6,10 @@
 }
 
 body {
-	height: 100vh;
-	display: flex;
-	align-items: center;
-	justify-content: center;
+	// height: 100vh;
+	// display: flex;
+	// align-items: center;
+	// justify-content: center;
 	font-family: "Overpass", sans-serif;
 }
